criação do eventBridge com auditoria e inicio da implementação dos alarmes com cloudWatch

+++ lambda/invoices/invoiceImportFunction.ts
@@ -1,5 +1,5 @@
 import { Context, S3Event, S3EventRecord } from "aws-lambda"
-import { ApiGatewayManagementApi, DynamoDB, S3 } from "aws-sdk"
+import { ApiGatewayManagementApi, EventBridge, DynamoDB, S3 } from "aws-sdk"
 import * as AWSRay from "aws-xray-sdk"
 import { InvoiceTransactionRepository, InvoiceTransactionStatus } from "/opt/nodejs/invoiceTransaction"
 import { InvoiceWSService } from "/opt/nodejs/invoiceWSConnection"
@@ -9,6 +9,9 @@ AWSRay.captureAWS(require('aws-sdk'))
 
 const invoicesDdb = process.env.INVOICE_DDB!
 const invoicesWsApiEndpoint = process.env.INVOICE_WSAPI_ENDPOINT!.substring(6)
+const auditBusName = process.env.AUDIT_BUS_NAME!
+const eventBridgeClient = new EventBridge();
+
 
 const s3Client = new S3();
 const ddbClient = new DynamoDB.DocumentClient();
@@ -58,28 +61,53 @@ async function processRecord(record: S3EventRecord) {
         const invoice = JSON.parse(object.Body!.toString('utf-8')) as InvoiceFile
         console.log(invoice)
 
-        const createInvoicePromise = invoiceRepository.create({
-            pk: `#invoice_${invoice.customerName}`,
-            sk: invoice.invoiceNumber,
-            ttl: 0,
-            totalValue: invoice.totalValue,
-            productId: invoice.productId,
-            quantity: invoice.quantity,
-            transactionId: key,
-            createdAt: Date.now()
-        })
+        if (invoice.invoiceNumber.length >= 5) {
+            const createInvoicePromise = invoiceRepository.create({
+                pk: `#invoice_${invoice.customerName}`,
+                sk: invoice.invoiceNumber,
+                ttl: 0,
+                totalValue: invoice.totalValue,
+                productId: invoice.productId,
+                quantity: invoice.quantity,
+                transactionId: key,
+                createdAt: Date.now()
+            })
 
-        const deleteObjectPromise = s3Client.deleteObject({
-            Key: key,
-            Bucket: record.s3.bucket.name
-        }).promise()
+            const deleteObjectPromise = s3Client.deleteObject({
+                Key: key,
+                Bucket: record.s3.bucket.name
+            }).promise()
 
-       const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED)
+            const puEventPromise = eventBridgeClient.putEvents({
+                Entries: [{
+                    Source: 'app.invoice',
+                    EventBusName: auditBusName,
+                    DetailType: 'invoice',
+                    Time: new Date(),
+                    Detail: JSON.stringify({
+                        errorDetails: 'FAIL_NO_INVOICE_NUMBER',
+                        info: {
+                            invoiceKey: key,
+                            customName: invoice.customerName,
+                        }
+                    })
+                }]
+            }).promise()
 
-        const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
+            const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED)
 
-        await Promise.all([createInvoicePromise, deleteObjectPromise, updateInvoicePromise, sendStatusPromise])
+            const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
 
+            await Promise.all([createInvoicePromise, deleteObjectPromise, updateInvoicePromise, sendStatusPromise, puEventPromise])
+        } else {
+            console.error(`invoice import failed - non valid invoice number - TransactionId: ${key}`)
+
+
+            const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
+            const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
+
+            await Promise.all([sendStatusPromise, updateInvoicePromise])
+        }
     } catch (error) {
         console.log((<Error>error).message)
     }
